legacy-python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pixel-dither-boy"
version = "1.0.0"
description = "Layered image-processing engine: ASCII, Dither, CRT, Glitch, Pixel Sort, Edge Detect"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "HRDWRE", email = "hello@hrdwre.dev" }]
keywords = ["image-processing", "glitch-art", "ascii", "dithering", "pixel-sort", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "streamlit>=1.32.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "requests>=2.31.0",
]

[project.scripts]
pixel-dither-boy = "scripts.cli:main"

[project.urls]
Homepage = "https://github.com/hrdwre/pixel-dither-boy"
Issues   = "https://github.com/hrdwre/pixel-dither-boy/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
